refactor(db): Extract shared lens utilities into lens_utils.rs
Phase 4 of the database.rs refactoring plan. Consolidate duplicated
lens/migration helper functions shared between LensedDocFetcher and
LensedAutoCommitFetcher.

New shared module `lens_utils.rs` (206 lines):
- doc_to_lens_doc: Convert Document to LensDoc
- lens_doc_to_doc: Convert LensDoc back to Document
- doc_needs_migration: Check if document needs migration
- versions_have_migrations: Check if any version has transforms
- build_collection_history: Build targeted history graph

Impact:
- lensed_auto_commit_fetcher.rs: 904 -> 753 lines (-151)
- lensed_fetcher.rs: 731 -> 637 lines (-94)
- Total reduction: ~245 lines of duplicated code removed
